gmaps_scraper/google_cloud_functions/requirements.txt
protobuf==3.19.1
PyYAML==6.0
